requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-heroku==0.3.1
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.1
requests==2.7.0
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
